lib: Refactor to allow model parsing to examine paged objects.
spatials calls can now provide a :paged option. In this case, the
parsing code is called multiple times, once for each page of objects.
Therefore the parsing code, if it chooses, doesn't have to care about
pages at all. It can be pretend it is dealing with a single page of
objects.

pdf_view.rb is modified to render object bounding boxes to multiple
pages.