sections.rb: New approach to section / header detection.
Aiming towards a new scheme that will be able to more accurately detect
sections and also detect headers and sub headers.

Regions are joined into sections where their font type and line heights
match and they flow on from each other according to detected columns.
Regions that are wider than the column width, or more than one line but
narrower than 90% of the column width are ignored. Adjacent regions
with the same line height and font are merged producing an array of
regions with varying line heights:

[x y z y z x z]

where z are potential section bodies, y the inner most header and x the
outer most.

The most common line height in terms of characters is calculated, and
is assumed to be a content body line height. The maximum distance from
each region with a non-body line height is calculated. The common line
height with the largest distance indicates the line height that
outer-most headings will have.