columns.rb: New column detection scheme.
New column detection scheme that has scope for better performance than
the old scheme. Bands or rows of the body area are checked for incidence
with characters, and those characters that do fall within the row are
appended to an x mask.

In the old scheme a few single-point lines were used to detect incidence
with regions when creating x masks for column detection. This will no
longer work now we generate columns from character positions rather than
region positons (in the hopes of using colunn information to form
characters into regions.)