test/test.rb: New testing system for pdf-extract.
Testing system that compares old, recorded reference extraction output
against current extraction output for a set of PDFs. If expected and new
output do not match the PDF is marked as failed.

Should provide a useful mechanism for keeping track of reference
extraction output changes as the project's extraction is refactored to
define regions in terms of zonal boundaries.